+++ okitweb/okitOci.py
@@ -197,9 +197,9 @@ def ociCompartment():
     logger.debug('Using Profile : {0!s:s}'.format(config_profile))
     oci_tenancies = OCITenancies(profile=config_profile)
     tenancy = oci_tenancies.listCompartments()
-    compartments = [{'display_name': c['display_name'], 'id': c['id'], 'home_region_key': tenancy['home_region_key']} for c in tenancy['compartments']]
-    compartments.append({'display_name': '/', 'id': tenancy['id'], 'home_region_key': tenancy['home_region_key']})
-    compartments.sort(key=lambda x: x['display_name'])
+    compartments = [{'display_name': c['display_name'], 'canonical_name': c['canonical_name'], 'id': c['id'], 'home_region_key': tenancy['home_region_key']} for c in tenancy['compartments']]
+    compartments.append({'display_name': '/', 'canonical_name': '/', 'id': tenancy['id'], 'home_region_key': tenancy['home_region_key']})
+    compartments.sort(key=lambda x: x['canonical_name'])
     logger.debug("Compartments: {0!s:s}".format(compartments))
     return json.dumps(compartments, sort_keys=False, indent=2, separators=(',', ': '))