+++ okitweb/static/okit/js/okit_console.js
@@ -4,8 +4,8 @@
 */
 console.info('Loaded Console Javascript');
 
-const okitVersion = '0.21.0';
-const okitReleaseDate = '5th May 2021';
+const okitVersion = '0.22.1';
+const okitReleaseDate = '26th May 2021';
 // Validation
 const validate_error_colour = "#ff4d4d";
 const validate_warning_colour = "#ffd633";