+++ okitweb/static/okit/model/js/artefacts/virtual_cloud_network.js
@@ -17,15 +17,33 @@ class VirtualCloudNetwork extends OkitArtifact {
         this.display_name = this.generateDefaultName(okitjson.virtual_cloud_networks.length + 1);
         this.compartment_id = data.parent_id;
         // Generate Cidr
-        this.cidr_block = '';
+        this.cidr_blocks = [''];
         this.dns_label = this.display_name.toLowerCase().slice(-6);
         this.is_ipv6enabled = false;
-        this.ipv6cidr_block = '';
+        this.ipv6cidr_blocks = [''];
         // Update with any passed data
         this.merge(data);
         this.convert();
+        Object.defineProperty(this, 'cidr_block', {get: function() {return this.cidr_blocks[0];}, set: function(cidr) {this.cidr_blocks[0] = cidr;}, enumerable: false });
+
     }
 
+    /*
+    ** Conversion Routine allowing loading of old json
+     */
+    convert() {
+        super.convert();
+        if (this.cidr_block !== undefined && this.cidr_block !== '') {
+            if (this.cidr_blocks === undefined || this.cidr_blocks.length === 0 || (this.cidr_blocks.length === 1 && this.cidr_blocks[0] === '')) this.cidr_blocks = [this.cidr_block];
+            else if (this.cidr_blocks.indexOf(this.cidr_block) < 0) this.cidr_blocks.push(this.cidr_block);
+            delete this.cidr_block;
+        }
+        if (this.ipv6cidr_block !== undefined && this.ipv6cidr_block !== '') {
+            if (this.ipv6cidr_blocks === undefined || this.ipv6cidr_blocks.length === 0 || (this.ipv6cidr_blocks.length === 1 && this.ipv6cidr_blocks[0] === '')) this.ipv6cidr_blocks = [this.ipv6cidr_block];
+            else if (this.ipv6cidr_blocks.indexOf(this.ipv6cidr_block) < 0) this.ipv6cidr_blocks.push(this.ipv6cidr_block);
+            delete this.ipv6cidr_block;
+        }
+    }
 
     /*
     ** Clone Functionality
@@ -219,7 +237,7 @@ class VirtualCloudNetwork extends OkitArtifact {
         let vcn_cidrs = [];
         for (let vcn of this.getOkitJson().getVirtualCloudNetworks()) {
             if (this.id !== vcn.id) {
-                vcn_cidrs.push(vcn.cidr_block.split('/')[0]);
+                vcn.cidr_blocks.forEach((cidr_block) => {vcn_cidrs.push(cidr_block.split('/')[0])});
             }
         }
         let second_octet = 0;
@@ -229,8 +247,8 @@ class VirtualCloudNetwork extends OkitArtifact {
             second_octet += 1;
         } while (vcn_cidrs.includes(ip));
 
-        this.cidr_block = `${ip}/16`;
-        return this.cidr_block;
+        this.cidr_blocks = [`${ip}/16`];
+        return this.cidr_blocks;
     }
 
     /*