+++ visualiser/discovery/models/__init__.py
@@ -6,9 +6,11 @@ from .core import ExtendedSecurityRule, ExtendedNetworkSecurityGroupVnic
 from .file_storage import ExtendedExportSummary
 from .object_storage import ExtendedPreauthenticatedRequestSummary
 from .mysql import ExtendedMySQLBackup, ExtendedMySQLBackupSummary
+from .database import ExtendedDbNodeSummary
 
 __all__ = [
   "ExtendedAutoScalingPolicySummary",
+  "ExtendedDbNodeSummary",
   "ExtendedExportSummary",
   "ExtendedMySQLBackup",
   "ExtendedMySQLBackupSummary",