+++ visualiser/generators/okitGenerator.py
@@ -57,7 +57,7 @@ class OCIGenerator(object):
         # -- Add Standard Author / Copyright variables
         self.jinja2_variables["author"] = __author__
         self.jinja2_variables["copyright"] = __copyright__
-        self.jinja2_variables["okit_version"] = "0.21.0"
+        self.jinja2_variables["okit_version"] = "0.22.1"
 
     def get(self, artifact_type, id):
         artifact = {};
@@ -170,18 +170,17 @@ class OCIGenerator(object):
         for network_security_group in self.visualiser_json.get('network_security_groups', []):
             self.renderNetworkSecurityGroup(network_security_group)
         # -- Security Lists
-        index = 1
         for security_list in self.visualiser_json.get('security_lists', []):
-            self.renderSecurityList(security_list, index)
-            index += 1
+            self.renderSecurityList(security_list)
         # -- Route Tables
-        index = 1
         for route_table in self.visualiser_json.get('route_tables', []):
-            self.renderRouteTable(route_table, index)
-            index += 1
+            self.renderRouteTable(route_table)
         # -- Service Gateways
         for service_gateway in self.visualiser_json.get('service_gateways', []):
             self.renderServiceGateway(service_gateway)
+        # -- Dhcp Options
+        for resource in self.visualiser_json.get('dhcp_options', []):
+            self.renderDhcpOption(resource)
         # -- Subnet
         for subnet in self.visualiser_json.get('subnets', []):
             self.renderSubnet(subnet)
@@ -552,6 +551,62 @@ class OCIGenerator(object):
 
         # -- Render Template
         jinja2_template = self.jinja2_environment.get_template("db_node.jinja2")
+        self.create_sequence.append(jinja2_template.render(self.jinja2_variables))
+        logger.debug(self.create_sequence[-1])
+        return
+
+    def renderDhcpOption(self, resource):
+        # Reset Variables
+        self.initialiseJinja2Variables()
+        # Read Data
+        standardisedName = self.standardiseResourceName(resource.get('resource_name', resource['display_name']))
+        resourceName = '{0:s}'.format(standardisedName)
+        self.jinja2_variables['resource_name'] = resourceName
+        self.jinja2_variables['output_name'] = resource['display_name']
+        # Process Route Table Data
+        logger.info('Processing Dhcp Options Information {0!s:s}'.format(standardisedName))
+        # -- Define Variables
+        # --- Required
+        # ---- Check if Default Dhcp Options
+        if resource.get("default", False): # Is it valid to replace the default DHCP Options
+            self.jinja2_variables["manage_default_resource_id"] = self.formatJinja2IdReference(
+                self.standardiseResourceName(self.id_name_map[resource['vcn_id']]), 'default_dhcp_options_id')
+        else:
+            self.removeJinja2Variable("manage_default_resource_id")
+        # ---- Compartment Id
+        self.jinja2_variables["compartment_id"] = self.formatJinja2IdReference(self.standardiseResourceName(self.id_name_map[resource['compartment_id']]))
+        # ---- Virtual Cloud Network OCID
+        self.jinja2_variables["vcn_id"] = self.formatJinja2IdReference(self.standardiseResourceName(self.id_name_map[resource['vcn_id']]))
+        # ---- Display Name
+        self.addJinja2Variable("display_name", resource["display_name"], standardisedName)
+        # ---- Route Rules
+        option_cnt = 1
+        jinja2_dhcp_options = []
+        for dhcp_option in resource.get('options', []):
+            jinja2_dhcp_option = {}
+            # ------ Type
+            logger.info(f'DHCP Option: {dhcp_option}')
+            jinja2_dhcp_option["type"] = self.generateJinja2Variable('dhcp_option_{0:02d}_type'.format(option_cnt), dhcp_option["type"], standardisedName)
+            # ------ Server Type
+            if dhcp_option["server_type"] != '':
+                jinja2_dhcp_option["server_type"] = self.generateJinja2Variable('dhcp_option_{0:02d}_server_type'.format(option_cnt), dhcp_option["server_type"], standardisedName)
+            # ------ Custom DNS Servers
+            if dhcp_option["custom_dns_servers"] != '':
+                jinja2_dhcp_option["custom_dns_servers"] = self.generateJinja2Variable('dhcp_option_{0:02d}_custom_dns_servers'.format(option_cnt), dhcp_option["custom_dns_servers"], standardisedName)
+            # ------ Domain Names
+            if dhcp_option["search_domain_names"] != '':
+                jinja2_dhcp_option["search_domain_names"] = self.generateJinja2Variable('dhcp_option_{0:02d}_search_domain_names'.format(option_cnt), dhcp_option["search_domain_names"], standardisedName)
+            # Add to Dhpc Option used for Jinja template
+            jinja2_dhcp_options.append(jinja2_dhcp_option)
+            # Increment option number
+            option_cnt += 1
+        self.jinja2_variables["options"] = jinja2_dhcp_options
+        # --- Optional
+        # ---- Tags
+        self.renderTags(resource)
+
+        # -- Render Template
+        jinja2_template = self.jinja2_environment.get_template("dhcp_option.jinja2")
         self.create_sequence.append(jinja2_template.render(self.jinja2_variables))
         logger.debug(self.create_sequence[-1])
         return
@@ -1389,7 +1444,7 @@ class OCIGenerator(object):
         logger.debug(self.create_sequence[-1])
         return
 
-    def renderRouteTable(self, route_table, index=0):
+    def renderRouteTable(self, route_table):
         # Reset Variables
         self.initialiseJinja2Variables()
         # Read Data
@@ -1401,8 +1456,8 @@ class OCIGenerator(object):
         logger.info('Processing Route Table Information {0!s:s}'.format(standardisedName))
         # -- Define Variables
         # --- Required
-        # ---- If this is the first Route Table the update the Default Route Table
-        if index == 1:
+        # ---- Check if Default Route Table
+        if route_table.get("default", False):
             self.jinja2_variables["manage_default_resource_id"] = self.formatJinja2IdReference(
                 self.standardiseResourceName(self.id_name_map[route_table['vcn_id']]), 'default_route_table_id')
         else:
@@ -1442,7 +1497,7 @@ class OCIGenerator(object):
         logger.debug(self.create_sequence[-1])
         return
 
-    def renderSecurityList(self, security_list, index=0):
+    def renderSecurityList(self, security_list):
         # Reset Variables
         self.initialiseJinja2Variables()
         # Read Data
@@ -1454,8 +1509,8 @@ class OCIGenerator(object):
         logger.info('Processing Security List Information {0!s:s}'.format(standardisedName))
         # -- Define Variables
         # --- Required
-        # ---- If this is the first Security List the update the Default Security List
-        if index == 1:
+        # ---- Check if Default Security List
+        if security_list.get("default", False):
             self.jinja2_variables["manage_default_resource_id"] = self.formatJinja2IdReference(
                 self.standardiseResourceName(self.id_name_map[security_list['vcn_id']]), 'default_security_list_id')
         else:
@@ -1652,8 +1707,8 @@ class OCIGenerator(object):
         else:
             self.jinja2_variables.pop("security_list_ids", None)
         # ---- DHCP Options
-        if 'dhcp_options' in subnet:
-            self.jinja2_variables["dhcp_options_id"] = self.formatJinja2DhcpReference(self.standardiseResourceName(subnet['dhcp_options']))
+        if 'dhcp_options_id' in subnet and len(subnet['dhcp_options_id']):
+            self.jinja2_variables["dhcp_options_id"] = self.formatJinja2DhcpReference(self.standardiseResourceName(subnet['dhcp_options_id']))
         else:
             self.jinja2_variables["dhcp_options_id"] = self.formatJinja2DhcpReference(self.standardiseResourceName(self.id_name_map[subnet['vcn_id']]))
         # --- Optional
@@ -1696,8 +1751,10 @@ class OCIGenerator(object):
         # --- Required
         # ---- Compartment Id
         self.jinja2_variables["compartment_id"] = self.formatJinja2IdReference(self.standardiseResourceName(self.id_name_map[virtual_cloud_network['compartment_id']]))
-        # ---- CIDR Block
-        self.addJinja2Variable("cidr_block", virtual_cloud_network["cidr_block"], standardisedName)
+        # ---- CIDR Blocks
+        # self.addJinja2Variable("cidr_blocks", virtual_cloud_network.get("cidr_blocks", [virtual_cloud_network.get("cidr_block", '')]), standardisedName)
+        self.jinja2_variables["cidr_blocks"] = virtual_cloud_network.get("cidr_blocks", [virtual_cloud_network.get("cidr_block", '')])
+        # self.addJinja2Variable("cidr_blocks", ",".join(virtual_cloud_network.get("cidr_blocks", [virtual_cloud_network.get("cidr_block", '')])), standardisedName)
         # ---- Display Name
         self.addJinja2Variable("display_name", virtual_cloud_network["display_name"], standardisedName)
         # ---- DNS Label
@@ -1708,7 +1765,8 @@ class OCIGenerator(object):
             # ----- Enabled
             self.addJinja2Variable("is_ipv6enabled", virtual_cloud_network["is_ipv6enabled"], standardisedName)
             # ----- Block
-            self.addJinja2Variable("ipv6cidr_block", virtual_cloud_network["ipv6cidr_block"], standardisedName)
+            # self.addJinja2Variable("ipv6cidr_block", virtual_cloud_network["ipv6cidr_block"], standardisedName)
+            self.jinja2_variables["ipv6cidr_blocks"] = virtual_cloud_network.get("ipv6cidr_blocks", [virtual_cloud_network.get("ipv6cidr_block", '')])
         else:
             self.jinja2_variables.pop("is_ipv6enabled", None)
             self.jinja2_variables.pop("ipv6cidr_block", None)